backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
pydantic>=2.5.0
anthropic>=0.39.0
google-genai>=0.3.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
